more in-depth module refactoring first pass
