+++ lib/cluster/moderator/modules/elasticsearch.js
@@ -3,8 +3,8 @@ var getClient = require('../../../utils/config').getClient;
 var Promise = require('bluebird');
 var elasticError = require('../../../utils/error_utils').elasticError;
 var _ = require('lodash');
-var nodeInfo = require('../../storage/backends/elasticsearch').nodeInfo;
-var nodeStats = require('../../storage/backends/elasticsearch').nodeStats;
+var nodeInfo = require('../../../data_sources/elasticsearch').nodeInfo;
+var nodeStats = require('../../../data_sources/elasticsearch').nodeStats;
 
 module.exports = function(context, logger) {
     var limit = context.sysconfig.teraslice.moderator_limit;
