+++ lib/cluster/runners/op.js
@@ -2,8 +2,10 @@
 
 var path = require('path');
 var exceptions = require('../../utils/exceptions');
-var walk = require('../../utils/file_utils').walk;
 var existsSync = require('../../utils/file_utils').existsSync;
+var fs = require('fs');
+var path = require('path');
+
 /*
  * This module defines the job execution context on the worker nodes.
  */
@@ -14,6 +16,18 @@ module.exports = function(context) {
     if (config.ops_directory) opPath = config.ops_directory;
 
     var types = ['readers', 'processors'];
+
+    function walk(rootDir, callback) {
+        fs.readdirSync(rootDir).forEach(function(filename) {
+            var filePath = path.join(rootDir, filename);
+            if (fs.statSync(filePath).isDirectory()) {
+                walk(filePath, callback);
+            } else {
+                callback(filePath, rootDir, filename);
+            }
+        });
+    }
+
 
     function findOp(name) {
         var filePath;
