+++ lib/cluster/services/jobs.js
@@ -4,7 +4,6 @@ var _ = require('lodash');
 var Promise = require('bluebird');
 
 var Queue = require('../../utils/queue');
-var exceptions = require('../../utils/exceptions');
 
 // Queue of jobs pending processing
 var pendingExecutionQueue;
@@ -333,10 +332,10 @@ module.exports = function(context, cluster_service) {
         return getExecutionContext(ex_id)
             .then(function(ex) {
                 if (ex._status === 'completed') {
-                    throw new exceptions.JobStateError("This job has completed and can not be restarted.");
+                    throw new Error("This job has completed and can not be restarted.");
                 }
                 if (ex._status === 'scheduling') {
-                    throw new exceptions.JobStateError("This job is currently being scheduled and can not be restarted.");
+                    throw new Error("This job is currently being scheduled and can not be restarted.");
                 }
                 // port cleanup and hostname happens at jobAllocator
                 ex._recover_execution = true;
