+++ lib/cluster/storage/backends/elasticsearch_store.js
@@ -21,7 +21,7 @@ module.exports = function(context, index_name, record_type, id_field, bulk_size)
     var bulkSize = 500;
     if (bulk_size) bulkSize = bulk_size;
 
-    function esGet(record_id) {
+    function get(record_id) {
         logger.trace(`getting record id: ${record_id}`);
         let query = {
             index: index_name,
@@ -32,7 +32,7 @@ module.exports = function(context, index_name, record_type, id_field, bulk_size)
         return elaticsearch.get(query)
     }
 
-    function esSearch(query, from, size, sort) {
+    function search(query, from, size, sort) {
         var esQuery = {
             index: index_name,
             from: from,
@@ -54,7 +54,7 @@ module.exports = function(context, index_name, record_type, id_field, bulk_size)
      * index saves a record to elasticsearch allowing automatic
      * ID creation
      */
-    function esIndex(record) {
+    function index(record) {
         logger.trace(`indexing record`, record);
         let query = {
             index: index_name,
@@ -70,7 +70,7 @@ module.exports = function(context, index_name, record_type, id_field, bulk_size)
      * index saves a record to elasticsearch with a specified ID.
      * If the document is already there it will be replaced.
      */
-    function esIndexWithId(record_id, record) {
+    function indexWithId(record_id, record) {
         logger.trace(`indexWithId call with id: ${record_id}, record`, record);
         let query = {
             index: index_name,
@@ -87,7 +87,7 @@ module.exports = function(context, index_name, record_type, id_field, bulk_size)
      * Create saves a record to elasticsearch under the provided id.
      * If the record already exists it will not be inserted.
      */
-    function esCreate(record) {
+    function create(record) {
         logger.trace('creating record', record);
 
         let query = {
@@ -101,7 +101,7 @@ module.exports = function(context, index_name, record_type, id_field, bulk_size)
         return elaticsearch.create(query);
     }
 
-    function esCount(query, from, sort) {
+    function count(query, from, sort) {
         var esQuery = {
             index: index_name,
             from: from,
@@ -119,7 +119,7 @@ module.exports = function(context, index_name, record_type, id_field, bulk_size)
         return elaticsearch.search(esQuery)
     }
 
-    function esUpdate(record_id, update_spec) {
+    function update(record_id, update_spec) {
         logger.trace(`updating record ${record_id}, `, update_spec);
 
         let query = {
@@ -136,7 +136,7 @@ module.exports = function(context, index_name, record_type, id_field, bulk_size)
         return elaticsearch.update(query)
     }
 
-    function esRemove(record_id) {
+    function remove(record_id) {
         logger.trace(`removing record ${record_id}`);
         let query = {
             index: index_name,
@@ -317,17 +317,17 @@ module.exports = function(context, index_name, record_type, id_field, bulk_size)
     }, 10000);
 
     var api = {
-        get: esGet,
-        search: esSearch,
+        get: get,
+        search: search,
         refresh: refresh,
-        index: esIndex,
-        indexWithId: esIndexWithId,
-        create: esCreate,
-        update: esUpdate,
+        index: index,
+        indexWithId: indexWithId,
+        create: create,
+        update: update,
         bulk: bulk,
-        remove: esRemove,
+        remove: remove,
         shutdown: shutdown,
-        count: esCount
+        count: count
     };
 
     return new Promise(function(resolve, reject) {
@@ -335,7 +335,7 @@ module.exports = function(context, index_name, record_type, id_field, bulk_size)
         var clientName = JSON.stringify(config.state);
 
         client = getClient(context, config.state, 'elasticsearch');
-        elaticsearch = require('./elasticsearch')(client, logger, null);
+        elaticsearch = require('./../../../data_sources/elasticsearch')(client, logger, null);
         return _createIndex()
             .then(function(results) {
                 return isAvailable();
