+++ lib/cluster/storage/logs.js
@@ -3,7 +3,7 @@
 var Promise = require('bluebird');
 var _ = require('lodash');
 var getClient = require('../../utils/config').getClient;
-var putTemplate = require('./backends/elasticsearch').putTemplate;
+var putTemplate = require('./../../data_sources/elasticsearch').putTemplate;
 
 module.exports = function(context) {
     if (_.includes(context.sysconfig.terafoundation.logging, 'elasticsearch')) {
