+++ lib/data_sources/elasticsearch.js
@@ -4,17 +4,25 @@ var uuid = require('uuid');
 var Promise = require('bluebird');
 var _ = require('lodash');
 var moment = require('moment');
-var filterResponse = require('../../../utils/elastic_utils').filterResponse;
-var elasticError = require('../../../utils/error_utils').elasticError;
+var dateMath = require('datemath-parser');
+var filterResponse = require('../utils/elastic_utils').filterResponse;
+var elasticError = require('../utils/error_utils').elasticError;
 
 // Module to manage persistence in Elasticsearch.
 // All functions in this module return promises that must be resolved to get the final result.
+
+//
+
+//TODO bring out utils code, query builders etc
+//TODO logs => store => data_sources
+// TODO fix autodates
+// TODO elasticError fix, if it needs to be a module
+
 
 module.exports = function(client, logger, opConfig) {
     let config = opConfig ? opConfig : {};
 
     function search(query) {
-        let config = opConfig ? opConfig : {};
         let retryTimer = {start: 5000, limit: 10000};
         let isCounting = query.size === 0;
 
@@ -90,7 +98,7 @@ module.exports = function(client, logger, opConfig) {
         let retryTimer = {start: 5000, limit: 10000};
 
         return new Promise(function(resolve, reject) {
-            var errHandler = errorHandler(retryTimer, getRecord, reject, logger);
+            var errHandler = _errorHandler(retryTimer, getRecord, reject, logger);
 
             function getRecord() {
                 client.get(query)
@@ -108,7 +116,7 @@ module.exports = function(client, logger, opConfig) {
         let retryTimer = {start: 5000, limit: 10000};
 
         return new Promise(function(resolve, reject) {
-            var errHandler = errorHandler(retryTimer, indexRecord, reject, logger);
+            var errHandler = _errorHandler(retryTimer, indexRecord, reject, logger);
 
             function indexRecord() {
                 client.index(query)
@@ -126,7 +134,7 @@ module.exports = function(client, logger, opConfig) {
         let retryTimer = {start: 5000, limit: 10000};
 
         return new Promise(function(resolve, reject) {
-            var errHandler = errorHandler(retryTimer, indexRecordID, reject, logger);
+            var errHandler = _errorHandler(retryTimer, indexRecordID, reject, logger);
 
             function indexRecordID() {
                 client.index(query)
@@ -144,7 +152,7 @@ module.exports = function(client, logger, opConfig) {
         let retryTimer = {start: 5000, limit: 10000};
 
         return new Promise(function(resolve, reject) {
-            var errHandler = errorHandler(retryTimer, createRecord, reject, logger);
+            var errHandler = _errorHandler(retryTimer, createRecord, reject, logger);
 
             function createRecord() {
                 client.create(query)
@@ -163,7 +171,7 @@ module.exports = function(client, logger, opConfig) {
         let retryTimer = {start: 5000, limit: 10000};
 
         return new Promise(function(resolve, reject) {
-            var errHandler = errorHandler(retryTimer, updateRecord, reject, logger);
+            var errHandler = _errorHandler(retryTimer, updateRecord, reject, logger);
 
             function updateRecord() {
                 client.update(query)
@@ -181,7 +189,7 @@ module.exports = function(client, logger, opConfig) {
         let retryTimer = {start: 5000, limit: 10000};
 
         return new Promise(function(resolve, reject) {
-            var errHandler = errorHandler(retryTimer, removeRecord, reject, logger);
+            var errHandler = _errorHandler(retryTimer, removeRecord, reject, logger);
 
             function removeRecord() {
                 client.delete(query)
@@ -194,9 +202,7 @@ module.exports = function(client, logger, opConfig) {
             removeRecord();
         });
     }
-
-//utility slicer
-
+    
     function parseDate(date) {
         var result;
 
@@ -212,7 +218,7 @@ module.exports = function(client, logger, opConfig) {
     }
 
 
-    function errorHandler(retryTimer, fn, reject, logger) {
+    function _errorHandler(retryTimer, fn, reject, logger) {
         return function(err) {
             if (_.get(err, 'body.error.type') === 'es_rejected_execution_exception') {
                 let timer = Math.floor(Math.random() * (retryTimer.limit - retryTimer.start) + retryTimer.start);
@@ -269,7 +275,7 @@ module.exports = function(client, logger, opConfig) {
             });
     }
 
-    function checkVersion(str) {
+    function _checkVersion(str) {
         var num = Number(str.replace(/\./g, ''));
         return num >= 210;
     }
@@ -280,7 +286,7 @@ module.exports = function(client, logger, opConfig) {
             .then(function(data) {
                 var version = data.nodes.versions[0];
 
-                if (checkVersion(version)) {
+                if (_checkVersion(version)) {
                     return client.indices.getSettings({})
                         .then(function(results) {
                             if (results[opConfig.index]) {
@@ -361,6 +367,55 @@ module.exports = function(client, logger, opConfig) {
     }
 
 
+    function _buildRangeQuery(source, obj) {
+        var date_field_name = source.date_field_name;
+        var dateObj = {};
+
+        dateObj[date_field_name] = {
+            gte: obj.start,
+            lt: obj.end
+        };
+
+        var body;
+
+        if (obj.key) {
+            body = {
+                query: {
+                    bool: {
+                        must: [
+                            {range: dateObj},
+                            {wildcard: {_uid: obj.key}}
+                        ]
+                    }
+                }
+            }
+        }
+        else {
+            body = {
+                query: {
+                    range: dateObj
+                }
+            };
+        }
+
+        return body;
+    }
+
+    function buildQuery(source, msg) {
+
+        var query = {
+            index: source.index,
+            size: msg.count,
+            body: _buildRangeQuery(source, msg)
+        };
+
+        if (source.query) {
+            query.q = source.query;
+        }
+
+        return query;
+    }
+
     return {
         search: search,
         get: get,
@@ -374,6 +429,8 @@ module.exports = function(client, logger, opConfig) {
         putTemplate: putTemplate,
         bulkSend: bulkSend,
         nodeInfo: nodeInfo,
-        nodeStats: nodeStats
+        nodeStats: nodeStats,
+        buildQuery: buildQuery,
+        parseDate: parseDate
     };
 };
