+++ lib/processors/elasticsearch_bulk.js
@@ -81,7 +81,7 @@ function newProcessor(context, opConfig, jobConfig) {
 
     function send(client, data) {
         var warning = warn(logger, 'The elasticsearch cluster queues are overloaded, resubmitting failed queries from bulk');
-        var elasticsearch = require('../cluster/storage/backends/elasticsearch')(client, logger, opConfig);
+        var elasticsearch = require('../data_sources/elasticsearch')(client, logger, opConfig);
         return elasticsearch.bulkSend(data, warning);
     }
