+++ lib/processors/elasticsearch_index_selector.js
@@ -1,11 +1,37 @@
 'use strict';
 
 var _ = require('lodash');
-var indexName = require('./../utils/processor_utils').indexName;
 var getOpConfig = require('../utils/config').getOpConfig;
 
+
 function newProcessor(context, opConfig, jobConfig) {
-    var opConfig = opConfig;
+
+    function formattedDate(record) {
+        var offsets = {
+            "daily": 10,
+            "monthly": 7,
+            "yearly": 4
+        };
+
+        var end = offsets[opConfig.timeseries] || 10;
+        var date = new Date(record[opConfig.date_field]).toISOString().slice(0, end);
+
+        return date.replace(/-/gi, '.');
+    }
+
+    function indexName(record) {
+        if (opConfig.timeseries) {
+
+            var index = formattedDate(record);
+            var prefix = opConfig.index_prefix;
+            var index_prefix = prefix.charAt(prefix.length - 1) === '-' ? prefix : prefix + '-';
+
+            return index_prefix + index;
+        }
+        else {
+            return opConfig.index
+        }
+    }
 
     // index_prefix is require if timeseries
     if (opConfig.timeseries && !opConfig.index_prefix) {
@@ -47,7 +73,7 @@ function newProcessor(context, opConfig, jobConfig) {
             var indexSpec = {};
 
             var meta = {
-                "_index": indexName(record, opConfig),
+                "_index": indexName(record),
                 "_type": opConfig.type ? opConfig.type : data.hits.hits[start]._type
             };
