+++ lib/readers/elasticsearch_data_generator.js
@@ -1,10 +1,33 @@
 'use strict';
 
-var parsedSchema = require('../utils/elastic_utils').parsedSchema;
 var Promise = require('bluebird');
 var getOpConfig = require('../utils/config').getOpConfig;
 var mocker = require('mocker-data-generator');
+var dataSchema = require('../utils/data_utils');
 
+function parsedSchema(opConfig) {
+    var schema = false;
+
+    if (opConfig.json_schema) {
+        var firstPath = opConfig.json_schema;
+        var nextPath = process.cwd() + '/' + opConfig.json_schema;
+
+        try {
+            if (fs.existsSync(firstPath)) {
+                schema = require(firstPath);
+            }
+            else {
+                schema = require(nextPath);
+            }
+        }
+        catch (e) {
+            throw new Error('Could not retrieve code for: ' + opConfig + '\n' + e);
+        }
+    }
+    else {
+        return dataSchema(opConfig, schema);
+    }
+}
 
 function newReader(context, opConfig, jobConfig) {
     var schema = parsedSchema(opConfig);
