+++ lib/readers/elasticsearch_date_range/reader.js
@@ -1,11 +1,9 @@
 'use strict';
 
-var buildQuery = require('./../../utils/elastic_utils').buildQuery;
-
 function newReader(context, opConfig, jobConfig, client) {
     return function(msg, logger) {
-        var query = buildQuery(opConfig, msg);
-        var elasticsearch = require('../../cluster/storage/backends/elasticsearch')(client, logger, opConfig);
+        var elasticsearch = require('../../data_sources/elasticsearch')(client, logger, opConfig);
+        var query = elasticsearch.buildQuery(opConfig, msg);
         return elasticsearch.search(query)
     }
 }
