+++ lib/readers/elasticsearch_date_range/slicer.js
@@ -2,10 +2,7 @@
 
 var Promise = require('bluebird');
 var moment = require('moment');
-var dateMath = require('datemath-parser');
 var _ = require('lodash');
-var processInterval = require('./../../utils/elastic_utils').processInterval;
-var buildQuery = require('./../../utils/elastic_utils').buildQuery;
 var dateOptions = require('./../../utils/elastic_utils').dateOptions;
 var dateFormatMS = require('./../../utils/elastic_utils').dateFormat;
 var dateFormatS = require('./../../utils/elastic_utils').dateFormatSeconds;
@@ -20,8 +17,8 @@ function newSlicer(context, opConfig, job, retryData, logger, client) {
     var time_resolution = dateOptions(opConfig.time_resolution);
 
     var dateFormat = time_resolution === 'ms' ? dateFormatMS : dateFormatS;
-    
-    var elasticsearch = require('../../cluster/storage/backends/elasticsearch')(client, logger, opConfig);
+
+    var elasticsearch = require('../../data_sources/elasticsearch')(client, logger, opConfig);
 
     function checkElasticsearch(client) {
         if (client) {
@@ -32,19 +29,43 @@ function newSlicer(context, opConfig, job, retryData, logger, client) {
         }
     }
 
+    function processInterval(str, esDates, opConfig) {
+        if (!moment(new Date(str)).isValid()) {
+            //one or more digits, followed by one or more letters, case-insensitive
+            var regex = /(\d+)([a-z]+)/i;
+            var interval = regex.exec(str);
 
-    function parseDate(date) {
-        var result;
+            if (interval === null) {
+                throw  new Error('elasticsearch_reader interval and/or delay are incorrectly formatted. Needs to follow ' +
+                    '[number][letter\'s] format, e.g. "12s"')
+            }
 
-        if (moment(new Date(date)).isValid()) {
-            result = moment(new Date(date));
+            //dont need first parameter, its the full string
+            interval.shift();
+            interval[1] = dateOptions(interval[1]);
+
+            return compareInterval(interval, esDates, opConfig);
         }
         else {
-            var ms = dateMath.parse(date);
-            result = moment(ms)
+            throw  new Error('elasticsearch_reader interval and/or delay are incorrectly formatted. Needs to follow ' +
+                '[number][letter\'s] format, e.g. "12s"')
+        }
+    }
+
+    function compareInterval(interval, esDates, opConfig) {
+        if (esDates) {
+            var datesDiff = esDates.limit.diff(esDates.start);
+            var intervalDiff = moment.duration(Number(interval[0]), interval[1]).as('milliseconds');
+
+            if (intervalDiff > datesDiff) {
+                if (opConfig.time_resolution === 's') {
+                    return [datesDiff / 1000, 's']
+                }
+                return [datesDiff, 'ms']
+            }
         }
 
-        return result;
+        return interval;
     }
 
     function getIndexDate(opConfig, client, date, order) {
@@ -54,8 +75,8 @@ function newSlicer(context, opConfig, job, retryData, logger, client) {
         var query = null;
 
         if (date) {
-            givenDate = parseDate(date);
-            query = buildQuery(opConfig, {count: 1, start: opConfig.start, end: opConfig.end})
+            givenDate = elasticsearch.parseDate(date);
+            query = elasticsearch.buildQuery(opConfig, {count: 1, start: opConfig.start, end: opConfig.end})
         }
         else {
             var sortOrder = order === 'start' ? 'asc' : 'desc';
@@ -435,7 +456,7 @@ function newSlicer(context, opConfig, job, retryData, logger, client) {
             range.key = key;
         }
 
-        var query = buildQuery(opConfig, range);
+        var query = elasticsearch.buildQuery(opConfig, range);
 
         return elasticsearch.search(query);
     }
