+++ lib/readers/id_reader.js
@@ -4,7 +4,6 @@ var _ = require('lodash');
 var getClient = require('../utils/config').getClient;
 var getOpConfig = require('../utils/config').getOpConfig;
 
-
 var parallelSlicers = true;
 
 function newSlicer(context, job, retryData, slicerAnalytics, logger) {
@@ -30,7 +29,7 @@ function newReader(context, opConfig, jobConfig) {
                 }
             }
         };
-        var elasticsearch = require('../cluster/storage/backends/elasticsearch')(client, opConfig, logger);
+        var elasticsearch = require('../data_sources/elasticsearch')(client, opConfig, logger);
         return elasticsearch.search(query);
     };
 }
