+++ lib/readers/id_slicer.js
@@ -2,8 +2,14 @@
 
 var _ = require('lodash');
 var elasticError = require('../utils/error_utils').elasticError;
-var getKeyArray = require('../utils/id_utils').getKeyArray;
-var buildQuery = require('../utils/elastic_utils').buildQuery;
+
+var base64url = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w',
+    'x', 'y', 'z', 'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X',
+    'Y', 'Z', '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', '\-', '_'];
+
+var hexadecimal = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'a', 'b', 'c', 'd', 'e', 'f'];
+
+var HEXADECIMAL = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'A', 'B', 'C', 'D', 'E', 'F'];
 
 
 module.exports = function(client, job, opConfig, logger, retryData, range) {
@@ -11,10 +17,22 @@ module.exports = function(client, job, opConfig, logger, retryData, range) {
     var keyRange = opConfig.key_range;
     var baseKeyArray = getKeyArray(opConfig);
     var keyArray = opConfig.key_range ? opConfig.key_range : baseKeyArray.slice();
-    var elasticsearch = require('../cluster/storage/backends/elasticsearch')(client, logger, opConfig);
-    
+    var elasticsearch = require('../data_sources/elasticsearch')(client, logger, opConfig);
+
     function getCountForKey(query) {
         return elasticsearch.search(query);
+    }
+
+    function getKeyArray(opConfig) {
+        if (opConfig.key_type === 'base64url') {
+            return base64url.slice();
+        }
+        if (opConfig.key_type === 'hexadecimal') {
+            return hexadecimal.slice();
+        }
+        if (opConfig.key_type === 'HEXADECIMAL') {
+            return HEXADECIMAL.slice();
+        }
     }
 
     function determineKeySlice(generator, closePath, rangeObj) {
@@ -50,7 +68,7 @@ module.exports = function(client, job, opConfig, logger, retryData, range) {
                 end: rangeObj.end,
                 key: key
             };
-            query = buildQuery(opConfig, range)
+            query = elasticsearch.buildQuery(opConfig, range)
         }
 
         return getCountForKey(query)
