+++ lib/utils/cluster.js
@@ -1,8 +1,5 @@
 'use strict';
 
-var _ = require('lodash');
-//var makeLogName = require('./config').makeLogName;
-
 function createNameSpace(job) {
     var jobName = job.name ? job.name : 'job';
     jobName = jobName.replace(' ', '').trim();
