+++ lib/utils/config.js
@@ -3,36 +3,6 @@
 var emitter = require('./events');
 
 
-function compareDates(prev, accum) {
-    var firstDate;
-    var secondDate;
-
-    if (accum.end) {
-        secondDate = new Date(accum.end)
-    }
-
-    if (prev.end) {
-        firstDate = new Date(prev.end);
-    }
-
-    if (secondDate && firstDate) {
-        if (secondDate > firstDate) {
-            return accum;
-        }
-        else {
-            return prev;
-        }
-    }
-
-    if (secondDate && !firstDate) {
-        return accum;
-    }
-
-    if (!secondDate && firstDate) {
-        return prev;
-    }
-}
-
 function getOpConfig(job, name) {
     return job.operations.find(function(op) {
         return op._op === name;
@@ -65,6 +35,5 @@ function getClient(context, config, type) {
 
 module.exports = {
     getClient: getClient,
-    compareDates: compareDates,
     getOpConfig: getOpConfig
 };
