+++ lib/utils/elastic_utils.js
@@ -3,7 +3,6 @@
 var Promise = require('bluebird');
 var moment = require('moment');
 var fs = require('fs');
-var dataSchema = require('./data_utils');
 var _ = require('lodash');
 var elasticError = require('./error_utils').elasticError;
 
@@ -24,118 +23,6 @@ function dateOptions(value) {
     }
     else {
         throw new Error('date interval is not formatted correctly')
-    }
-}
-
-function processInterval(str, esDates, opConfig) {
-    if (!moment(new Date(str)).isValid()) {
-        //one or more digits, followed by one or more letters, case-insensitive
-        var regex = /(\d+)([a-z]+)/i;
-        var interval = regex.exec(str);
-
-        if (interval === null) {
-            throw  new Error('elasticsearch_reader interval and/or delay are incorrectly formatted. Needs to follow ' +
-                '[number][letter\'s] format, e.g. "12s"')
-        }
-
-        //dont need first parameter, its the full string
-        interval.shift();
-        interval[1] = dateOptions(interval[1]);
-
-        return compareInterval(interval, esDates, opConfig);
-    }
-    else {
-        throw  new Error('elasticsearch_reader interval and/or delay are incorrectly formatted. Needs to follow ' +
-            '[number][letter\'s] format, e.g. "12s"')
-    }
-}
-
-function compareInterval(interval, esDates, opConfig) {
-    if (esDates) {
-        var datesDiff = esDates.limit.diff(esDates.start);
-        var intervalDiff = moment.duration(Number(interval[0]), interval[1]).as('milliseconds');
-
-        if (intervalDiff > datesDiff) {
-            if (opConfig.time_resolution === 's') {
-                return [datesDiff / 1000, 's']
-            }
-            return [datesDiff, 'ms']
-        }
-    }
-
-    return interval;
-}
-
-function buildRangeQuery(source, obj) {
-    var date_field_name = source.date_field_name;
-    var dateObj = {};
-
-    dateObj[date_field_name] = {
-        gte: obj.start,
-        lt: obj.end
-    };
-
-    var body;
-
-    if (obj.key) {
-        body = {
-            query: {
-                bool: {
-                    must: [
-                        {range: dateObj},
-                        {wildcard: {_uid: obj.key}}
-                    ]
-                }
-            }
-        }
-    }
-    else {
-        body = {
-            query: {
-                range: dateObj
-            }
-        };
-    }
-
-    return body;
-}
-
-function buildQuery(source, msg) {
-
-    var query = {
-        index: source.index,
-        size: msg.count,
-        body: buildRangeQuery(source, msg)
-    };
-
-    if (source.query) {
-        query.q = source.query;
-    }
-
-    return query;
-}
-
-function parsedSchema(opConfig) {
-    var schema = false;
-
-    if (opConfig.json_schema) {
-        var firstPath = opConfig.json_schema;
-        var nextPath = process.cwd() + '/' + opConfig.json_schema;
-
-        try {
-            if (fs.existsSync(firstPath)) {
-                schema = require(firstPath);
-            }
-            else {
-                schema = require(nextPath);
-            }
-        }
-        catch (e) {
-            throw new Error('Could not retrieve code for: ' + opConfig + '\n' + e);
-        }
-    }
-    else {
-        return dataSchema(opConfig, schema);
     }
 }
 
@@ -210,11 +97,6 @@ var dateFormatSeconds = "YYYY-MM-DDTHH:mm:ssZ";
 
 module.exports = {
     dateOptions: dateOptions,
-    processInterval: processInterval,
-    compareInterval: compareInterval,
-    buildRangeQuery: buildRangeQuery,
-    buildQuery: buildQuery,
-    parsedSchema: parsedSchema,
     dateFormat: dateFormat,
     dateFormatSeconds: dateFormatSeconds,
     logState: logState,
